refactor(bodyweights): migrate from drizzle to spring boot

+++ backend/src/main/java/com/liftts/backend/domain/dtos/CreateBodyweightRequestDto.java
@@ -19,7 +19,7 @@ public class CreateBodyweightRequestDto {
 
     @NotNull(message = "Weight is required")
     @Min(value = 1, message = "Weight should be greater than or equal to {value}kg/lb")
-    private Integer weight;
+    private Double weight;
 
     @NotNull(message = "Bodyweight date is required")
     private LocalDate date;
